+++ main.tf
@@ -57,6 +57,17 @@ resource "aws_s3_bucket" "content" {
   bucket        = local.content_bucket_name
   force_destroy = var.force_destroy_buckets
   tags          = merge(var.tags, var.tags_s3_bucket_content, { Name = "${var.domain_name} Static Content" })
+
+  dynamic "cors_rule" {
+    for_each = length(var.cors_allowed_origins) > 0 ? ["create"] : []
+    content {
+      allowed_headers = var.cors_allowed_headers
+      allowed_methods = var.cors_allowed_methods
+      allowed_origins = var.cors_allowed_origins
+      expose_headers  = []
+      max_age_seconds = 3000
+    }
+  }
 }
 data "aws_s3_bucket" "content" {
   count  = var.create_content_bucket ? 0 : 1