+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.3.0
 
 - Configurable cloudfront methods using `cloudfront_allowed_methods`.  Valid options are "get", "get_options", "all"
+- Configurable CORS rule when bucket is created by the module
 
 ## v3.2.0